A hospital Chatbot.
This was developed when covid-19 was rampant in India. I wanted to test myself and tried to create a chatbot which could search up available hospital beds closest to the person via whatsapp.

The code uses requests to scrape a particular website for beds and hospitals, while it uses selenium to push the whatsapp messages.